backend/pyproject.toml
[project]
name = "tictactoe-api"
version = "0.1.0"
description = "Professional Tic-Tac-Toe API with FastAPI"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "python-multipart>=0.0.7",
    "email-validator>=2.1.0",
    "passlib[bcrypt]>=1.7.4",
    "tenacity>=8.2.3",
    "pydantic>=2.0",
    "pydantic-settings>=2.2.1",
    "alembic>=1.12.1",
    "httpx>=0.25.1",
    "aiomysql>=0.2.0",
    "sqlmodel>=0.0.21",
    "bcrypt==4.0.1",
    "pyjwt>=2.8.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "mypy>=1.8.0",
    "ruff>=0.2.2",
    "pre-commit>=3.6.2",
    "types-passlib>=1.7.7",
    "coverage>=7.4.3",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]
line-length = 79

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e
]

[tool.black]
line-length = 79
target-version = ['py310', 'py311']

[tool.isort]
profile = "black"
line_length = 79

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=app --cov-report=html --cov-report=term-missing"
asyncio_mode = "auto"
